Intragrated all features on viewHoles Page. Search, filter by district and order by. Solved a bug regarding updating holes. Solved a bug on deleting a hole with null team. Solved a css problem regarding top-margin on viewHoles

+++ src/main/java/com/nicolrom/services/impl/TeamServiceImpl.java
@@ -55,7 +55,9 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public void deleteTeam(Team team) {
-        teamDao.deleteTeam(team);
+        if (team != null){
+            teamDao.deleteTeam(team);
+        }
     }
 
     private List<Integer> parseEmployeesStringArray(List<String> employeesStringArray){
